% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{install.packages}
\alias{install.packages}
\title{Install Packages (First Installation)}
\usage{
install.packages(...)
}
\arguments{
\item{packages}{pre-define packages list.}
}
\description{
To install required packages. Use ipak function to install desired packages.
}
\examples{
install.packages()
}
\keyword{install.pack}
