% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{step3}
\alias{step3}
\title{Step 3}
\usage{
step3(
  project_title = "Project Title",
  xpt_location = ".\\\\",
  prog_location = "\\\\programs\\\\",
  define_location = "\\\\datasets\\\\",
  range_character = "yes"
)
}
\arguments{
\item{logo}{Add logo by copying logo.jpg to the main working folderfirst time package user, run this function to install all required packages.
Can be downloaded from my Github}

\item{style}{Load document style by copying word template and rename as style.docx to the main working folder.
Can be downloaded from my Github}

\item{title}{Project title}
}
\description{
Convert csv to xpt and create define document
}
\examples{
step4(title="Add title here")
}
\keyword{step3}
