% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{lib.update}
\alias{lib.update}
\title{Update library}
\usage{
lib.update(
  prevlib =
    "//certara.com/sites/S02-Cary/Consulting/Projects/ras/Nath team/library/xptlibrary.csv",
  newlib = "studydefinelist.csv"
)
}
\description{
Optional used function.
}
\examples{
create.library()
}
\keyword{definelist}
